+++ src/Equinox.Projection/Projection.fs
@@ -1,3 +1,0 @@
-﻿namespace Equinox.Projection
-
-type [<NoComparison>] StreamItem = { stream: string; index: int64; event: Equinox.Codec.IEvent<byte[]> }